Create Palindrome Code
To solve this problem, we can use the following approach:

Check if the string is already a palindrome. If it is, return -1.
Use two pointers, one starting from the beginning and the other from the end of the string. Move the pointers towards each other until they point to different characters.
When the pointers point to different characters, check if removing the character at the left pointer or the character at the right pointer results in a palindrome.
Return the index of the character that needs to be removed to make the string a palindrome.

This code first checks if the string is already a palindrome. If not, it finds the index of the character that needs to be removed to make the string a palindrome.